Move beacon pool endpoints from http api to its own module.
