+++ beacon_node/http_api/src/beacon/mod.rs
@@ -1,1 +1,2 @@
+pub mod pool;
 pub mod states;
